requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
adversarial-robustness-toolbox=1.13.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
build=0.10.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h74dc2b5_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cleverhans=4.0.0=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
contourpy=1.0.5=py310hdb19cb5_0
cryptography=38.0.4=py310h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
eagerpy=0.30.0=pypi_0
easydict=1.10=pypi_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
foolbox=3.3.3=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.3.0=pypi_0
gast=0.5.3=pypi_0
giflib=5.2.1=h5eee18b_1
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lightning-utilities=0.8.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.6.2=py310h06a4308_0
matplotlib-base=3.6.2=py310h945d387_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mnist=0.2.2=pypi_0
multidict=6.0.4=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nose=1.3.7=pypi_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=22.0=py310h06a4308_0
path=16.6.0=pypi_0
path-py=12.5.0=pypi_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pillow=9.3.0=py310h6a678d5_2
pip=22.3.1=py310h06a4308_0
pip-tools=6.13.0=pypi_0
plotly=5.14.1=pypi_0
ply=3.11=py310h06a4308_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
pycodestyle=2.10.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310h06a4308_0
pyproject-hooks=1.0.0=pypi_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.6=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.12.0=py3.10_cuda10.2_cudnn7.6.5_0
pytorch-lightning=2.0.0=pypi_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.25.1=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.10.0=pypi_0
sentry-sdk=1.15.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
tenacity=8.2.2=pypi_0
tensorflow-probability=0.19.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pypi_0
torchattacks=3.4.0=pypi_0
torchaudio=0.12.0=py310_cu102
torchmetrics=0.11.4=pypi_0
torchvision=0.13.0=py310_cu102
tornado=6.2=py310h5eee18b_0
tqdm=4.64.1=py310h06a4308_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py310h06a4308_0
wandb=0.13.10=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
